rename plaza_processing files for shorter and clearer structure
