+++ plaza_preprocessing/plaza_preprocessing/__main__.py
@@ -1,10 +1,11 @@
 import logging
 import sys
 
-from plaza_preprocessing import osm_importer
-from plaza_preprocessing.osm_merger import osm_merger
-from plaza_preprocessing.osm_optimizer import osm_optimizer
-from plaza_preprocessing.osm_optimizer.graphprocessor.visibilitygraphprocessor import VisibilityGraphProcessor
+from plaza_preprocessing.importer import importer
+from plaza_preprocessing.merger import merger
+from plaza_preprocessing.optimizer import optimizer
+from plaza_preprocessing.optimizer.graphprocessor.visibilitygraph import VisibilityGraphProcessor
+from plaza_preprocessing.optimizer.graphprocessor.spiderwebgraph import SpiderWebGraphProcessor
 
 logger = logging.getLogger('plaza_preprocessing')
 
@@ -27,9 +28,9 @@ def setup_logging(verbose=False, quiet=False):
 
 
 def preprocess_osm(osm_filename, out_file):
-    osm_holder = osm_importer.import_osm(osm_filename)
-    processed_plazas = osm_optimizer.preprocess_plazas(osm_holder, VisibilityGraphProcessor())
-    osm_merger.merge_plaza_graphs(processed_plazas, osm_filename, out_file)
+    osm_holder = importer.import_osm(osm_filename)
+    processed_plazas = optimizer.preprocess_plazas(osm_holder, VisibilityGraphProcessor())
+    merger.merge_plaza_graphs(processed_plazas, osm_filename, out_file)
 
 
 if __name__ == "__main__":
