+++ plaza_preprocessing/plaza_preprocessing/importer/osmholder.py
@@ -1,0 +1,8 @@
+
+class OSMHolder:
+    """holder for importer OSM objects"""
+    def __init__(self, plazas, buildings, lines, points):
+        self.plazas = plazas
+        self.buildings = buildings
+        self.lines = lines
+        self.points = points
