+++ plaza_preprocessing/plaza_preprocessing/merger/geojson_writer.py
@@ -1,5 +1,6 @@
 import geojson
 
+
 def write_geojson(geometries, filename):
     """ write geojson from a list of shapely geometries """
     features = [geojson.Feature(geometry=geom) for geom in geometries]
