+++ plaza_preprocessing/plaza_preprocessing/optimizer/graphprocessor/spiderwebgraph.py
@@ -1,8 +1,8 @@
 from math import ceil
 from shapely.geometry import Point, LineString, Polygon
 from typing import List
-from plaza_preprocessing.osm_optimizer import utils
-from plaza_preprocessing.osm_optimizer.graphprocessor.graphprocessor import GraphProcessor
+from plaza_preprocessing.optimizer import utils
+from plaza_preprocessing.optimizer.graphprocessor.graphprocessor import GraphProcessor
 
 
 class SpiderWebGraphProcessor(GraphProcessor):
